Adding session settings to utilise toggles
