+++ background.html
@@ -60,6 +60,7 @@
       </div>
     </div>
     <div class='conversation-stack'>
+      <div id='settings-container'></div>
       <div class='conversation placeholder'>
         <div class='conversation-header'></div>
         <div class='container'>
@@ -700,6 +701,7 @@
   <script type='text/javascript' src='js/views/last_seen_indicator_view.js'></script>
   <script type='text/javascript' src='js/views/scroll_down_button_view.js'></script>
   <script type='text/javascript' src='js/views/toast_view.js'></script>
+  <script type='text/javascript' src='js/views/session_settings_view.js'></script>
   <script type='text/javascript' src='js/views/session_toast_view.js'></script>
   <script type='text/javascript' src='js/views/session_toggle_view.js'></script>
   <script type='text/javascript' src='js/views/session_modal_view.js'></script>
