+++ js/background.js
@@ -802,6 +802,17 @@
       appView.openConversation(groupId, {});
     };
 
+
+    
+    $(document).ready(() => {
+      window.settingsView = new Whisper.SessionSettingsView({
+        el: $('#settings-container'),
+      });
+      window.settingsView.render();
+    });
+
+
+    
     window.generateID = () =>
       Math.random()
         .toString(36)
