+++ stylesheets/_session.scss
@@ -687,3 +687,9 @@ label {
   margin-top: 50px;
   margin-left: 75px;
 }
+
+#settings-container {
+  width: 100%;
+  height: 100%;
+  background-color: blue;
+}
